Enhancements to the spark job enrolments and acbp
