+++ batch-models/src/main/scala/org/ekstep/analytics/dashboard/report/acbp/UserACBPReportModel.scala
@@ -40,31 +40,9 @@ object UserACBPReportModel extends AbsDashboardModel {
     // get ACBP details data frame
     val acbpDF = acbpDetailsDF()
 
-    // CustomUser
-    val acbpCustomUserAllotmentDF = acbpDF
-      .filter(col("assignmentType") === "CustomUser")
-      .withColumn("userID", explode(col("assignmentTypeInfo")))
-      .join(userDataDF, Seq("userID", "userOrgID"), "left")
-    show(acbpCustomUserAllotmentDF, "acbpCustomUserAllotmentDF")
-
-    // Designation
-    val acbpDesignationAllotmentDF = acbpDF
-      .filter(col("assignmentType") === "Designation")
-      .withColumn("designation", explode(col("assignmentTypeInfo")))
-      .join(userDataDF, Seq("userOrgID", "designation"), "left")
-    show(acbpDesignationAllotmentDF, "acbpDesignationAllotmentDF")
-
-    // All User
-    val acbpAllUserAllotmentDF = acbpDF
-      .filter(col("assignmentType") === "AllUser")
-      .join(userDataDF, Seq("userOrgID"), "left")
-    show(acbpAllUserAllotmentDF, "acbpAllUserAllotmentDF")
-
-    // union of all the response dfs
-    val acbpAllotmentDF = Seq(acbpCustomUserAllotmentDF, acbpDesignationAllotmentDF, acbpAllUserAllotmentDF).map(df => {
-      df.select("userID", "fullName", "userPrimaryEmail", "userMobile", "designation", "group", "userOrgID", "ministry_name", "dept_name", "userOrgName", "acbpID", "assignmentType", "completionDueDate", "allocatedOn", "acbpCourseIDList","acbpStatus", "acbpCreatedBy","cbPlanName")
-    }).reduce((a, b) => a.union(b))
-    show(acbpAllotmentDF, "acbpAllotmentDF")
+    val selectColumns = Seq("userID", "fullName", "userPrimaryEmail", "userMobile", "designation", "group", "userOrgID", "ministry_name", "dept_name", "userOrgName", "acbpID",
+      "assignmentType", "completionDueDate", "allocatedOn", "acbpCourseIDList","acbpStatus", "acbpCreatedBy","cbPlanName")
+    val acbpAllotmentDF = explodedACBPDetails(acbpDF, userDataDF, selectColumns)
 
     // replace content list with names of the courses instead of ids
     val acbpAllEnrolmentDF = acbpAllotmentDF
